RoomView: Move the search logic into a separate module.
